+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapBuilder.java
@@ -174,4 +174,9 @@ class GoogleMapBuilder implements GoogleMapOptionsSink {
   public void setInitialTileOverlays(List<Map<String, ?>> initialTileOverlays) {
     this.initialTileOverlays = initialTileOverlays;
   }
+
+  @Override
+  public void setMapId(String mapId) {
+    options.mapId(mapId);
+  }
 }