+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapController.java
@@ -848,6 +848,11 @@ final class GoogleMapController
     }
   }
 
+  @Override
+  public void setMapId(String mapId) {
+    Log.e(TAG, "Cannot change MapId after map is initialized.");
+  }
+
   private void updateInitialTileOverlays() {
     tileOverlaysController.addTileOverlays(initialTileOverlays);
   }