+++ packages/google_maps_flutter/google_maps_flutter_ios/example/integration_test/google_maps_test.dart
@@ -16,6 +16,7 @@ const LatLng _kInitialMapCenter = LatLng(0, 0);
 const double _kInitialZoomLevel = 5;
 const CameraPosition _kInitialCameraPosition =
     CameraPosition(target: _kInitialMapCenter, zoom: _kInitialZoomLevel);
+const String _kCloudMapId = '8e0a97af9386fef';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -1024,6 +1025,19 @@ void main() {
       expect(tileOverlayInfo1, isNull);
     },
   );
+
+  testWidgets('testSetStyleMapId', (WidgetTester tester) async {
+    final Key key = GlobalKey();
+
+    await tester.pumpWidget(Directionality(
+      textDirection: TextDirection.ltr,
+      child: ExampleGoogleMap(
+        key: key,
+        initialCameraPosition: _kInitialCameraPosition,
+        cloudMapId: _kCloudMapId,
+      ),
+    ));
+  });
 }
 
 class _DebugTileProvider implements TileProvider {