+++ packages/google_maps_flutter/google_maps_flutter_ios/example/lib/example_google_map.dart
@@ -247,6 +247,7 @@ class ExampleGoogleMap extends StatefulWidget {
     this.onCameraIdle,
     this.onTap,
     this.onLongPress,
+    this.cloudMapId,
   }) : super(key: key);
 
   /// Callback method for when the map is ready to be used.
@@ -348,6 +349,12 @@ class ExampleGoogleMap extends StatefulWidget {
 
   /// Which gestures should be consumed by the map.
   final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+
+  /// Identifier that's associated with a specific cloud bases map style.
+  ///
+  /// See https://developers.google.com/maps/documentation/get-map-id
+  /// for more details.
+  final String? cloudMapId;
 
   /// Creates a [State] for this [ExampleGoogleMap].
   @override
@@ -534,5 +541,6 @@ MapConfiguration _configurationFromMapWidget(ExampleGoogleMap map) {
     indoorViewEnabled: map.indoorViewEnabled,
     trafficEnabled: map.trafficEnabled,
     buildingsEnabled: map.buildingsEnabled,
+    cloudMapId: map.cloudMapId,
   );
 }