+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,4 +1,8 @@
-## NEXT
+## 2.4.0
+
+* Adds implementation for `cloudMapId` parameter to support cloud-based maps styling.
+
+## 2.3.0
 
 * Updates minimum Flutter version to 3.0.